Add files via upload
A simple credit card approval classification model code that provides the output whether an individual should be approved for the credit card and if so, what should be the credit limit to be provided. Assumption - Min Credit Limit = $1000, Max Credit Limit - $ 8000

The focus of the code is not to get high accuracy results.